state: Track shard and queue-id in headers from Django and Tornado.
By being an intermediary for the internal queue-creation request
between Django and Tornado, we can see what Tornado shard the Django
process is asking for, as well as what queue-id is generated.  We
track this association in an in-memory DashMap (a concurrent-access
hashmap) and use it to route later requests which come from the
clients, by way of nginx.

The resulting server is stateful, and must also be able to respond
with a 400 status code and `BAD_EVENT_QUEUE_ID` JSON message if the
queue-id is unknown.  While this ensures that its behaviour is
_correct_ if it is restarted, it will lead to every client
re-requesting a queue, with severe implications for load.  Future
commits will address this.