handler: Use X-Accel-Redirect to offload GET requests.
`X-Accel-Redirect` tells nginx to forward the request as a GET request
to some other route.  This offloads the handling of the longpoll to
nginx, which would have to handle *a* longpoll regardless.  We are
only able to do this with GET requests, so we must still proxy queue
creation and deletion personally.

This results in a fraction of a millisecond lower latency in no-op
non-blocking fetches against the fake-tornado backend when under heavy
load, and up to 50ms longer latencies under high load -- but a
significantly higher number of maximum connections that can be served
with the same CPU cycles when benchmarked with a fake-tornado backend.